+++ app/components/shared/__tests__/IconSidebar.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import IconSidebar from '@/app/components/shared/IconSidebar';
-import { SidebarProvider } from '@/app/context/SidebarContext';
-import { ThemeProvider } from '@/app/context/ThemeContext';
+import { SidebarProvider } from '@/app/providers/SidebarContext';
+import { ThemeProvider } from '@/app/providers/ThemeContext';
 
 const Wrapper = ({ children }: { children: React.ReactNode }) => (
   <ThemeProvider>
