+++ app/features/layout/context/HeaderVisibilityContext.tsx
