+++ app/features/surah/[surahId]/_components/ArabicFontPanel.tsx
@@ -2,10 +2,10 @@
 'use client';
 import { FaArrowLeft } from '@/app/components/shared/SvgIcons';
 import { useTranslation } from 'react-i18next';
-import { useSettings } from '@/app/context/SettingsContext';
+import { useSettings } from '@/app/providers/SettingsContext';
 import { useState } from 'react'; // Import useState
-import { useTheme } from '@/app/context/ThemeContext';
-import { useHeaderVisibility } from '@/app/context/HeaderVisibilityContext';
+import { useTheme } from '@/app/providers/ThemeContext';
+import { useHeaderVisibility } from '@/app/features/layout/context/HeaderVisibilityContext';
 
 interface ArabicFontPanelProps {
   isOpen: boolean;
