+++ app/features/surah/[surahId]/_components/Verse.tsx
@@ -12,7 +12,7 @@ import { Verse as VerseType, Translation, Word } from '@/types';
 import type { LanguageCode } from '@/lib/text/languageCodes';
 import { useAudio } from '@/app/features/player/context/AudioContext';
 import Spinner from '@/app/components/shared/Spinner';
-import { useSettings } from '@/app/context/SettingsContext';
+import { useSettings } from '@/app/providers/SettingsContext';
 import { applyTajweed } from '@/lib/text/tajweed';
 
 interface VerseProps {
