+++ app/features/surah/[surahId]/page.tsx
@@ -16,7 +16,7 @@ import {
 import { LANGUAGE_CODES } from '@/lib/languageCodes';
 import type { LanguageCode } from '@/lib/languageCodes';
 import { WORD_LANGUAGE_LABELS } from '@/lib/wordLanguages';
-import { useSettings } from '@/app/context/SettingsContext';
+import { useSettings } from '@/app/providers/SettingsContext';
 import Spinner from '@/app/components/shared/Spinner';
 import useSWR from 'swr';
 import useSWRInfinite from 'swr/infinite';
