+++ app/features/surah/__tests__/SurahPage.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
-import { SettingsProvider } from '@/app/context/SettingsContext';
+import { SettingsProvider } from '@/app/providers/SettingsContext';
 import { AudioProvider } from '@/app/features/player/context/AudioContext';
-import { ThemeProvider } from '@/app/context/ThemeContext';
-import { SidebarProvider } from '@/app/context/SidebarContext';
+import { ThemeProvider } from '@/app/providers/ThemeContext';
+import { SidebarProvider } from '@/app/providers/SidebarContext';
 import SurahPage from '@/app/features/surah/[surahId]/page';
 import { Verse } from '@/types';
 import * as api from '@/lib/api';
