+++ app/features/tafsir/[surahId]/[ayahId]/_components/VerseCard.tsx
@@ -9,7 +9,7 @@ import {
 import { Verse as VerseType, Translation, Word } from '@/types';
 import type { LanguageCode } from '@/lib/text/languageCodes';
 import { useAudio } from '@/app/features/player/context/AudioContext';
-import { useSettings } from '@/app/context/SettingsContext';
+import { useSettings } from '@/app/providers/SettingsContext';
 import Spinner from '@/app/components/shared/Spinner';
 import { applyTajweed } from '@/lib/text/tajweed';
 import DOMPurify from 'dompurify';
