+++ app/features/tafsir/__tests__/TafsirTabs.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import TafsirTabs from '@/app/features/tafsir/[surahId]/[ayahId]/_components/TafsirTabs';
-import { ThemeProvider } from '@/app/context/ThemeContext';
-import { SettingsProvider } from '@/app/context/SettingsContext';
+import { ThemeProvider } from '@/app/providers/ThemeContext';
+import { SettingsProvider } from '@/app/providers/SettingsContext';
 import { getTafsirCached } from '@/lib/tafsir/tafsirCache';
 import useSWR from 'swr';
